refactor(util): 수강 상태 확인 로직 및 시간 계산 유틸 함수 리팩토링
- 커리큘럼 섹션별 총 강의 시간 표시 추가
- 중복된 시간 계산 로직을 getTotalLectureDuration 유틸 함수로 분리하여 통합 관리
- 시간의 총합을 계산하는 로직 추가로 timeFormat 유틸함수 이름을 time으로 변경
